Move nock logic back to generated file
