+++ packages/cli/templates/actions/snapshot/__tests__/snapshot.test.ts
@@ -1,13 +1,32 @@
-import { createTestIntegration } from '@segment/actions-core'
+import { createTestEvent, createTestIntegration } from '@segment/actions-core'
+import { generateTestData } from '../../../lib/test-data'
 import destination from '../index'
 import prettier from 'prettier'
+import nock from 'nock'
 
 const testDestination = createTestIntegration(destination)
 
 describe('Testing snapshot for {{destination}} action:', () => {
   for (const actionSlug in destination.actions) {
     it(`${actionSlug} [required]`, async () => {
-      const responses = await testDestination.snapshotAction(destination, actionSlug, true)
+      const action = destination.actions[actionSlug]
+      const [eventData, settingsData] = generateTestData(destination, action, true)
+
+      nock(/.*/).persist().get(/.*/).reply(200)
+      nock(/.*/).persist().post(/.*/).reply(200)
+      nock(/.*/).persist().put(/.*/).reply(200)
+
+      const event = createTestEvent({
+        properties: eventData
+      })
+
+      const responses = await testDestination.testAction(actionSlug, {
+        event: event,
+        mapping: event.properties,
+        settings: settingsData,
+        auth: undefined
+      })
+
       const request = responses[0].request
 
       expect(request.headers).toMatchSnapshot()
@@ -23,7 +42,24 @@ describe('Testing snapshot for {{destination}} action:', () => {
     })
 
     it(`${actionSlug} [all fields]`, async () => {
-      const responses = await testDestination.snapshotAction(destination, actionSlug, false)
+      const action = destination.actions[actionSlug]
+      const [eventData, settingsData] = generateTestData(destination, action, false)
+
+      nock(/.*/).persist().get(/.*/).reply(200)
+      nock(/.*/).persist().post(/.*/).reply(200)
+      nock(/.*/).persist().put(/.*/).reply(200)
+
+      const event = createTestEvent({
+        properties: eventData
+      })
+
+      const responses = await testDestination.testAction(actionSlug, {
+        event: event,
+        mapping: event.properties,
+        settings: settingsData,
+        auth: undefined
+      })
+
       const request = responses[0].request
 
       if (request.body) {
